one code editor for each sprite
