+++ src/components/PyatchEditor.jsx
@@ -1,24 +1,17 @@
 import React, { useContext } from 'react';
 import pyatchContext from './provider/PyatchContext.js';
-import CodeMirror from '@uiw/react-codemirror';
-import { material } from '@uiw/codemirror-theme-material';
-import { python } from '@codemirror/lang-python';
+import { PyatchTargetEditor } from './PyatchTargetEditor.jsx';
 
-export function PyatchCodeEditor(props) {
-    const { pyatchEditor } = useContext(pyatchContext);
-    const { activeSprite } = useContext(pyatchContext);
+const PyatchEditor = () => {
+    let { sprites } = useContext(pyatchContext);
 
-    const updateState = (newValue) => {
-        pyatchEditor.editorText[activeSprite] = newValue;
-    }
-
-    return(
-        <CodeMirror
-            value={pyatchEditor.editorText[activeSprite]}
-            extensions={[python()]}
-            theme={material}
-            onChange={updateState}
-            height="90vh"
-        />
+    return (
+        <div>
+            {sprites.map((sprite) => {
+                return <PyatchTargetEditor spriteID={sprite}/>
+            })}
+        </div>
     );
 }
+
+export default PyatchEditor
